+++ sdk/azidentity/CHANGELOG.md
@@ -9,9 +9,11 @@
 ### Bugs Fixed
 
 ### Other Changes
-* Exported `CredentialUnavailableError`. Custom credentials should return
-  this type from their `GetToken()` methods when they can't acquire a token
-  and `ChainedTokenCredential` should try its next credential, if any.
+* `AzureCLICredential` imposes its default timeout only when the `Context`
+  passed to `GetToken()` has no deadline
+* Added `NewCredentialUnavailableError()`. This function constructs an error indicating
+  a credential can't authenticate and an encompassing `ChainedTokenCredential` should
+  try its next credential, if any.
 
 ## 1.3.0-beta.1 (2022-12-13)